Moved udfgenerator to mipengine module. This was made in order to remove the dependency of the controller from the node module. The way it was before the controller was importing the algorithms that were importing the node/udfgenerator module hence creating a dependency of the controller from the node module.
