+++ mipengine/algorithms/logistic_regression.py
@@ -1,13 +1,10 @@
 # type: ignore
-from numbers import Number
-from typing import DefaultDict
 from typing import TypeVar
-from typing import Any
 
 import numpy
 import pandas as pd
 
-from mipengine.node.udfgen import (
+from mipengine.udfgen import (
     TensorBinaryOp,
     TensorUnaryOp,
     literal,
